+++ src/Console/JsonProcessor.php
@@ -23,7 +23,7 @@ final class JsonProcessor implements Processor
         $path = null;
         foreach ($unitOfWorks as $unitOfWork) {
             $unitOfWork->run();
-            $path ??= $unitOfWork->source->getFileName();
+            $path ??= $unitOfWork->path;
         }
 
         $this->exporter->writeln(null === $path ? [] : ['path' => $path, 'data' => $unitOfWorks]);