+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All Notable changes to `bakame/stackwatch` will be documented in this file.
 
+## [Next - African city](https://github.com/bakame-php/stackwatch/compare/0.11.0...main) - TBD
+
+### Added
+
+- None
+
+### Fixed
+
+- **BC BREAK:** Changed `Bakama\Stachwatcher\Marker::summary` method name to `Bakama\Stachwatcher\Marker::summarize`
+- Internal improvement, Lazy evaluation is now done in the `UnitOfWork`
+
+### Deprecated
+
+- None
+
+### Removed
+
+- None
+
+
 ## [0.11.0 - Kampala](https://github.com/bakame-php/stackwatch/compare/0.10.0...0.11.0) - 2025-08-03
 
 ### Added